LayoutTests/imported/w3c/web-platform-tests/shadow-dom/reference-target/tentative/reference-target-basics-expected.txt


PASS ShadowRoot.referenceTarget defaults to empty string when shadow is created declaratively
PASS <template> shadowrootreferencetarget sets referenceTarget on shadow root
PASS ShadowRoot.referenceTarget defaults to empty string when shadow is created imperatively
PASS ShadowRootInitDict can be used to set referenceTarget on shadow root
